etl_project/requirements.txt
agate==1.7.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
babel==2.15.0
backoff==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
connectorx==0.3.3
croniter==2.0.5
cryptography==42.0.7
dagster==1.7.6
dagster-cloud==1.7.6
dagster-cloud-cli==1.7.6
dagster-dbt==0.23.6
dagster-graphql==1.7.5
dagster-pipes==1.7.6
dagster-postgres==0.23.6
dagster-webserver==1.7.5
dbt-core==1.7.14
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
docstring-parser==0.16
-e file:///Users/charlie/Documents/code/data_engineering/toptal/etl_project
faker==25.2.0
faker-commerce==1.0.4
filelock==3.14.0
fsspec==2024.3.1
github3-py==4.0.1
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.63.0
grpcio-health-checking==1.62.2
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.7
importlib-metadata==6.11.0
iniconfig==2.0.0
isodate==0.6.1
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
leather==0.4.0
logbook==1.5.3
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==2.1.5
mashumaro==3.13
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parsedatetime==2.6
pathspec==0.11.2
pendulum==3.0.0
pex==2.3.1
pluggy==1.5.0
polars==0.20.26
prompt-toolkit==3.0.36
protobuf==4.25.3
psycopg2-binary==2.9.9
py4j==0.10.9.7
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.18.0
pyjwt==2.8.0
pyspark==3.5.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
pyyaml==6.0.1
questionary==2.0.1
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.30
sqlglot==23.17.0
sqlglotrs==0.2.5
sqlparse==0.5.0
starlette==0.37.2
structlog==24.1.0
tabulate==0.9.0
text-unidecode==1.3
time-machine==2.14.1
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
universal-pathlib==0.2.2
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.18.2
